reorganized files, removed unused imports
