+++ sim/app/w/[id]/components/panel/components/console/components/json-view/json-view.tsx
@@ -1,5 +1,5 @@
 import { useEffect, useState } from 'react'
-import { Download, Image as ImageIcon } from 'lucide-react'
+import { Download } from 'lucide-react'
 import { Button } from '@/components/ui/button'
 
 interface JSONViewProps {
