+++ sim/app/w/[id]/components/workflow-block/components/sub-block/components/eval-input.tsx
@@ -1,4 +1,3 @@
-import { useRef, useState } from 'react'
 import { Plus, Trash } from 'lucide-react'
 import { Button } from '@/components/ui/button'
 import { Input } from '@/components/ui/input'
