+++ sim/app/w/[id]/components/workflow-block/components/sub-block/components/file-upload.tsx
@@ -1,10 +1,9 @@
 'use client'
 
 import { useRef, useState } from 'react'
-import { Info, Upload, X } from 'lucide-react'
+import { X } from 'lucide-react'
 import { Button } from '@/components/ui/button'
 import { Progress } from '@/components/ui/progress'
-import { Tooltip, TooltipContent, TooltipTrigger } from '@/components/ui/tooltip'
 import { useNotificationStore } from '@/stores/notifications/store'
 import { useWorkflowRegistry } from '@/stores/workflows/registry/store'
 import { useSubBlockStore } from '@/stores/workflows/subblock/store'
