+++ sim/app/w/[id]/components/workflow-block/components/sub-block/components/schedule/schedule-config.tsx
@@ -8,7 +8,6 @@ import { formatDateTime } from '@/lib/utils'
 import { getWorkflowWithValues } from '@/stores/workflows'
 import { useWorkflowRegistry } from '@/stores/workflows/registry/store'
 import { useSubBlockStore } from '@/stores/workflows/subblock/store'
-import { workflowSync } from '@/stores/workflows/sync'
 import { useWorkflowStore } from '@/stores/workflows/workflow/store'
 import { useSubBlockValue } from '../../hooks/use-sub-block-value'
 import { ScheduleModal } from './components/schedule-modal'
