+++ sim/app/w/components/sidebar/components/settings-modal/components/api-keys/api-keys.tsx
@@ -1,7 +1,7 @@
 'use client'
 
 import { useEffect, useState } from 'react'
-import { AlertCircle, Check, Copy, KeySquare, Plus, Trash2 } from 'lucide-react'
+import { Check, Copy, KeySquare, Plus, Trash2 } from 'lucide-react'
 import {
   AlertDialog,
   AlertDialogAction,
@@ -27,7 +27,6 @@ import { Label } from '@/components/ui/label'
 import { Skeleton } from '@/components/ui/skeleton'
 import { useSession } from '@/lib/auth-client'
 import { createLogger } from '@/lib/logs/console-logger'
-import { cn } from '@/lib/utils'
 
 const logger = createLogger('ApiKeys')
