+++ sim/app/w/components/sidebar/sidebar.tsx
@@ -4,14 +4,13 @@ import { useMemo, useState } from 'react'
 import Link from 'next/link'
 import { usePathname, useRouter } from 'next/navigation'
 import clsx from 'clsx'
-import { HelpCircle, Plus, ScrollText, Settings, Store } from 'lucide-react'
+import { HelpCircle, Plus, ScrollText, Settings } from 'lucide-react'
 import { AgentIcon } from '@/components/icons'
 import { Button } from '@/components/ui/button'
 import { Tooltip, TooltipContent, TooltipTrigger } from '@/components/ui/tooltip'
 import { useWorkflowRegistry } from '@/stores/workflows/registry/store'
 import { WorkflowMetadata } from '@/stores/workflows/registry/types'
 import { HelpModal } from './components/help-modal/help-modal'
-import { NavItem } from './components/nav-item/nav-item'
 import { SettingsModal } from './components/settings-modal/settings-modal'
 import { WorkflowFolders } from './components/workflow-folders/workflow-folders'
