+++ sim/stores/custom-tools/store.ts
@@ -1,39 +1,6 @@
 import { create } from 'zustand'
 import { devtools, persist } from 'zustand/middleware'
-
-export interface CustomToolSchema {
-  type: string
-  function: {
-    name: string
-    description?: string
-    parameters: {
-      type: string
-      properties: Record<string, any>
-      required?: string[]
-    }
-  }
-}
-
-export interface CustomToolDefinition {
-  id: string
-  title: string
-  schema: CustomToolSchema
-  code: string
-  createdAt: string
-  updatedAt?: string
-}
-
-interface CustomToolsStore {
-  tools: Record<string, CustomToolDefinition>
-  addTool: (tool: Omit<CustomToolDefinition, 'id' | 'createdAt' | 'updatedAt'>) => string
-  updateTool: (
-    id: string,
-    updates: Partial<Omit<CustomToolDefinition, 'id' | 'createdAt' | 'updatedAt'>>
-  ) => boolean
-  removeTool: (id: string) => void
-  getTool: (id: string) => CustomToolDefinition | undefined
-  getAllTools: () => CustomToolDefinition[]
-}
+import { CustomToolsStore } from './types'
 
 export const useCustomToolsStore = create<CustomToolsStore>()(
   devtools(
