+++ sim/stores/execution/types.ts
@@ -1,0 +1,30 @@
+import { Executor } from '@/executor'
+import { ExecutionContext } from '@/executor/types'
+
+export interface ExecutionState {
+  activeBlockIds: Set<string>
+  isExecuting: boolean
+  isDebugging: boolean
+  pendingBlocks: string[]
+  executor: Executor | null
+  debugContext: ExecutionContext | null
+}
+
+export interface ExecutionActions {
+  setActiveBlocks: (blockIds: Set<string>) => void
+  setIsExecuting: (isExecuting: boolean) => void
+  setIsDebugging: (isDebugging: boolean) => void
+  setPendingBlocks: (blockIds: string[]) => void
+  setExecutor: (executor: Executor | null) => void
+  setDebugContext: (context: ExecutionContext | null) => void
+  reset: () => void
+}
+
+export const initialState: ExecutionState = {
+  activeBlockIds: new Set(),
+  isExecuting: false,
+  isDebugging: false,
+  pendingBlocks: [],
+  executor: null,
+  debugContext: null,
+}
