+++ sim/stores/ollama/types.ts
@@ -1,0 +1,4 @@
+export interface OllamaStore {
+  models: string[]
+  setModels: (models: string[]) => void
+}
